# Optimize markers via new layers API and treap-based index