--- a/spec/sample.js
+++ b/spec/sample.js
@@ -1,1 +1,0 @@
-undefined