… package.json
… spec/async-spec-helpers.coffee
… spec/atom-environment-spec.coffee
… spec/display-buffer-spec.coffee
… spec/text-editor-component-spec.coffee
… spec/text-editor-component-spec.js
… spec/text-editor-presenter-spec.coffee
… spec/text-editor-spec.coffee
… spec/view-registry-spec.coffee
… src/text-editor-component.coffee
… src/text-editor-element.coffee
… src/text-editor-marker-layer.coffee
… src/text-editor-marker.coffee
… src/text-editor-presenter.coffee
… src/text-editor.coffee
… src/view-registry.coffee