This PR improves the responsiveness of find-and-replace and other marker-intensive tasks by adding a layer-oriented API for working with collections of related markers and upgrading to a new implementation of the marker index. Searching for `e` in `jquery.js` now blocks the UI thread for ~1 second instead of ~3.2 seconds. Inserting newlines in this file with the markers present now takes ~10 ms instead of ~35 ms.

I still need to make the changes to find-and-replace backwards compatible and merge them after merging these changes.

## API Changes

### Marker Layers

A huge problem with our existing implementation is that creating markers emits synchronous events to support existing packages and core APIs. We needed a way to opt out of these event handlers for high-volume use cases, and marker layers were a good way to do this.

* `TextBuffer`
  * `.addMarkerLayer({maintainHistory})` Creates and returns a `MarkerLayer` object. If `maintainHistory` is `true`, markers will be restored on calls to undo/redo.
  * `.getMarkerLayer` Retrieve a marker layer based on its id.

* `MarkerLayer` Some highlights:
  * `.markRange` Creates a marker.
  * `.getMarker` Gets a marker by id.
  * `.findMarkers` Queries the marker layer.
  * `.onDidUpdate` Called asynchronously whenever markers are created, updated, or destroyed. Can be used without being a drag on performance, but you'll have to re-query any state you're interested in between calls.
  * `.onDidCreateMarker` This is needed to support our current synchronous selection APIs, but it shouldn't be used on layers that will contain a large number of markers because it slows down insertion.
  * `.copy` Create and return a new marker layer with all the same markers.

* `TextEditor`
  * `.addMarkerLayer({maintainHistory})` Creates and returns a `TextEditorMarkerLayer`.
  * `.getMarkerLayer` Retrieve an editor-level marker layer based on its id.

* `TextEditorMarkerLayer` This is a wrapper around a buffer-level `MarkerLayer` object and shares the same id.
  * `.markBufferRange` Create a marker.
  * `.markScreenRange` Create a marker.
  * `.findMarkers` Query markers with the same API as `TextEditor.prototype.findMarkers`

### Marker Layer Decorations

Another huge source of slowness in find and replace is that we decorated each marker individually. Now that we have marker layers, we can decorate an entire layer in the same way, avoiding a ton of allocations.

* `TextEditor`
  * `.decorateMarkerLayer(layer, properties)` Accepts a `MarkerLayer` or `TextEditorMarkerLayer` and the same properties that `decorateMarker` takes. It creates a decoration for every marker in the layer based on these parameters. This won't work for custom gutter decorations or overlay decorations, but *will* work for highlights, line, and line number decorations. Returns a `LayerDecoration`.
* `LayerDecoration`
  * `setPropertiesForMarker(marker, properties)` This can be used to override the properties of a particular marker in the layer. We needed this to support the "current result" in find and replace without major changes to the package code. Pass `null` to clear the override.

## New Marker Index

![marker-binary-tree](https://cloud.githubusercontent.com/assets/1789/10924350/062ead24-823b-11e5-85ee-ffd1e77df82f.png)

This PR also incorporates a new binary-tree-based [marker-index](https://github.com/atom/marker-index), which replaces the previous b-tree-based index in text-buffer. It offers better and more predictable performance... about 33% faster on my insertion tests (from 1.6 second to 1 second) and an order of magnitude faster at splicing. 1 second is still too long to block the UI thread, so we may still need to background the index population. I would like to experiment with a C++ implementation and see what it buys us, since it's a straightforward implementation that should be fairly easy to port.

## Performance Improvements

Now on to the good part... the improvement. This shows results from 3 implementations. The current version, marker layers with the old index, and marker layers with the new index.

### Without marker layers (current implementation)

#### Inserting ~20k markers:

~3.2 seconds

<img width="1143" alt="screen shot 2015-11-03 at 1 20 25 pm" src="https://cloud.githubusercontent.com/assets/1789/10922006/b26d99d2-822d-11e5-87fa-ff24243d1b5d.png">

#### Inserting newlines

~35 ms per keystroke

<img width="1098" alt="screen shot 2015-11-03 at 1 20 11 pm" src="https://cloud.githubusercontent.com/assets/1789/10922023/c65e319a-822d-11e5-90e0-176d6016d2f7.png">

### Old marker index

#### Inserting ~20k markers:

1.6 seconds

<img width="1239" alt="screen shot 2015-11-03 at 11 48 37 am" src="https://cloud.githubusercontent.com/assets/1789/10919754/61b5d024-8221-11e5-9155-4b5f2f17bbcf.png">

#### Inserting newlines

~30 ms per keystroke

<img width="1229" alt="screen shot 2015-11-03 at 11 51 12 am" src="https://cloud.githubusercontent.com/assets/1789/10920772/deeb95ba-8226-11e5-8ffa-7c9497bd399e.png">

Zooming in, the newline insertion contains 3 ~5ms calls to `MarkerIndex.prototype.splice`:

<img width="1119" alt="screen shot 2015-11-03 at 12 21 51 pm" src="https://cloud.githubusercontent.com/assets/1789/10920873/633900a0-8227-11e5-9ab5-eedd594e6606.png">

#### Memory consumed

36.2 MB

<img width="777" alt="screen shot 2015-11-03 at 1 15 50 pm" src="https://cloud.githubusercontent.com/assets/1789/10921926/31b3e3fa-822d-11e5-9b00-1c48426ba481.png">

### New marker index

#### Inserting ~20k markers

~1 second

<img width="1232" alt="screen shot 2015-11-03 at 12 45 31 pm" src="https://cloud.githubusercontent.com/assets/1789/10921238/3a37de72-8229-11e5-9854-9e92223029e9.png">

#### Inserting newlines

~7-10 ms per keystroke

<img width="1010" alt="screen shot 2015-11-03 at 12 48 16 pm" src="https://cloud.githubusercontent.com/assets/1789/10921246/490c5342-8229-11e5-84c9-a77a58b685dc.png">

Only 1 significant splice operation, 0.8ms.

<img width="829" alt="screen shot 2015-11-03 at 12 52 55 pm" src="https://cloud.githubusercontent.com/assets/1789/10921321/d89cf606-8229-11e5-889a-c45636f52fac.png">

#### Memory consumed

34.8 MB (not a huge savings)

<img width="778" alt="screen shot 2015-11-03 at 1 13 42 pm" src="https://cloud.githubusercontent.com/assets/1789/10921946/4b5947a0-822d-11e5-984a-4a5402c41251.png">