@EasyHard 2015-11-05
Hi @nathansobo, I have a feeling that maybe we can use splay tree instead of treap since operations have locality. How do you think? If you think it may be a good idea. I would like to give it a shot.