@nathansobo 2015-11-06
> Hi @nathansobo, I have a feeling that maybe we can use splay tree instead of treap since operations have locality

@EasyHard I considered using splay trees early on, but my concern is that a splay tree might not deliver consistent performance. While it is true that the *amortized* cost of access is optimal, splay trees can get in really unbalanced states such that a given operation would take longer. For batch workloads, this isn't a problem, but since Atom is an interactive application it's important that all operations have a fairly uniform cost.

Consider the example of populating the marker index in find and replace... We'll insert a bunch of markers in order, which without rebalancing will create a reversed linked-list structure. Then when we query the list for the first time, we'll have to walk the entire list to get to the beginning of the list. At this point we'll rebalance, but this first operation will be expensive.

It may be the case that we'll always find a way to amortize the cost without users noticing, but this wasn't obvious enough to me to want to allow our access patterns to be the only thing balancing the tree. Let me know if you think I'm missing something, but I'd prefer to stick to something more predictable.