@nathansobo 2015-11-07
I converted `text-editor-component-spec.coffee` to Babel as part of this PR. Kind of off topic, but I was having trouble understanding asynchronous sequencing in that test after removing all the mocking, and I really wanted `async/await`. Turns out most of my issues were probably related to throttling, but the conversion was a ton of work so I'm keeping it.