@nathansobo 2015-11-08
Really wondering what's going on with the build. Worked out most of the flakiness but I'm still timing out on animation frames that never fire on CI. From what I can tell there *is* a pending frame when I time out, but for some reason it's not occurring.

Now testing with an animation loop that logs if there's a frame > 500ms. May throw in the towel and go back to mocking the event loop, but I want to spend a bit more time understanding what the heck is going on with the event loop.