@nathansobo 2015-11-09
Okay, I have at least one green build on our internal CI, which hasn't happened yet. So either I got lucky or have found the fix. Now I need to narrow down which of the many things I tried are necessary for stable passing builds.